Separate tests into three sections, use klona for copying VCs.
- Remove unused helper functions and deps.
