+++ package.json
@@ -30,21 +30,22 @@
   "dependencies": {
     "@digitalbazaar/did-method-key": "^2.0.0",
     "@digitalbazaar/ed25519-signature-2020": "^3.0.0",
-    "@digitalbazaar/ezcap": "^2.0.3",
     "@digitalbazaar/http-client": "^2.0.1",
     "@digitalbazaar/vc": "^2.1.0",
+    "@digitalbazaar/vc-status-list": "^2.1.0",
+    "@digitalbazaar/vc-status-list-context": "^2.0.0",
     "@digitalcredentials/did-context": "^1.0.0",
     "bnid": "^2.1.0",
     "chai": "^4.3.4",
     "credentials-context": "^2.0.0",
     "ed25519-signature-2020-context": "^1.1.0",
     "jsonld-document-loader": "^1.2.0",
+    "klona": "^2.0.5",
     "mocha": "^9.1.3",
     "mocha-w3c-interop-reporter": "github:digitalbazaar/mocha-w3c-interop-reporter",
     "require-dir": "^1.2.0",
-    "@digitalbazaar/vc-status-list": "^2.1.0",
-    "@digitalbazaar/vc-status-list-context": "^2.0.0",
-    "vc-api-test-suite-implementations": "github:digitalbazaar/vc-api-test-suite-implementations#update-db-implementation"
+    "uuid": "^8.3.2",
+    "vc-api-test-suite-implementations": "github:digitalbazaar/vc-api-test-suite-implementations#add-status-list-tag-verifier"
   },
   "devDependencies": {
     "eslint": "^7.23.0",
