+++ tests/helpers.js
@@ -3,16 +3,11 @@
  */
 'use strict';
 
-const {decodeSecretKeySeed} = require('bnid');
 const {decodeList} = require('@digitalbazaar/vc-status-list');
-const didKey = require('@digitalbazaar/did-method-key');
-const {Ed25519Signature2020} = require('@digitalbazaar/ed25519-signature-2020');
 const {httpClient} = require('@digitalbazaar/http-client');
 const https = require('https');
-const {ZcapClient} = require('@digitalbazaar/ezcap');
 
 const agent = new https.Agent({rejectUnauthorized: false});
-const didKeyDriver = didKey.driver();
 
 // Javascript's default ISO timestamp is contains milliseconds.
 // This lops off the MS part of the UTC RFC3339 TimeStamp and replaces
@@ -20,8 +15,6 @@ const didKeyDriver = didKey.driver();
 const ISOTimeStamp = ({date = new Date()} = {}) => {
   return date.toISOString().replace(/\.\d+Z$/, 'Z');
 };
-
-const deepClone = data => JSON.parse(JSON.stringify(data, null, 2));
 
 const getCredentialStatus = async ({verifiableCredential}) => {
   // get SLC for the VC
@@ -40,25 +33,29 @@ const getCredentialStatus = async ({verifiableCredential}) => {
   return {status, statusListCredential: revocationListCredential};
 };
 
-async function getZcapClient() {
-  if(!process.env.CLIENT_SECRET_DB) {
-    throw new Error('ENV variable CLIENT_SECRET_DB is required.');
+/**
+ * Takes in a Map and a predicate and returns a new Map
+ * only returning key value pairs that are true.
+ *
+ * @param {object} options - Options to use.
+ * @param {Map} options.map - A Map.
+ * @param {Function<boolean>} options.predicate - A function to
+ * filter the map's entries on.
+ *
+ * @returns {Map} Returns a map.
+ */
+const filterMap = ({map, predicate}) => {
+  const filtered = new Map();
+  for(const [key, value] of map) {
+    const result = predicate({key, value});
+    if(result === true) {
+      filtered.set(key, value);
+    }
   }
-  const secretKeySeed = process.env.CLIENT_SECRET_DB;
-  const seed = await decodeSecretKeySeed({secretKeySeed});
-  const didKey = await didKeyDriver.generate({seed});
-  const {didDocument: {capabilityInvocation}} = didKey;
-  const zcapClient = new ZcapClient({
-    SuiteClass: Ed25519Signature2020,
-    invocationSigner: didKey.keyPairs.get(capabilityInvocation[0]).signer(),
-    agent
-  });
-  return zcapClient;
-}
-
+  return filtered;
+};
 module.exports = {
   ISOTimeStamp,
-  deepClone,
-  getCredentialStatus,
-  getZcapClient
+  filterMap,
+  getCredentialStatus
 };
